requirements.txt
fastapi
uvicorn
email-validator
sqlmodel
python-dotenv
pymysql
psycopg2-binary